Undid stupid thing u64->u32

+++ deno_webgpu/device.rs
@@ -849,7 +849,7 @@ impl GPUDevice {
                     .into_iter()
                     .map(|attr| wgpu_types::VertexAttribute {
                       format: attr.format.into(),
-                      offset: attr.offset as u32,
+                      offset: attr.offset,
                       shader_location: attr.shader_location,
                     })
                     .collect(),
